Remove extra COOKIE header from the gmail profile
When using the gmail profile we found that the extra `COOKIE` header caused issues (ie. base64 decoding errors). My guess is that CS expects the first `COOKIE` value to be the value that contains magic content that needs to be base64-decoded.

Removing this extra header made the problem go away.